backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aci-sdk>=1.0.0b2",
    "elevenlabs>=2.4.0",
    "fastapi>=0.115.13",
    "openai>=1.90.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "rprint>=0.0.8",
    "uvicorn>=0.34.3",
]
